requirements.txt
google-api-python-client
google-auth-oauthlib
requests
dotenv
openai
ffmpeg-python